# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: A3 Partner Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/alerts/partner/{partnerId}:
    get:
      operationId: getPartnerAlerts
      tags: [admin]
      parameters: [{ name: partnerId, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Alerts, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
  /admin/alerts/order/{orderId}:
    get:
      operationId: getOrderAlerts
      tags: [admin]
      parameters: [{ name: orderId, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Alerts, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
  /admin/alerts/manual-followup/{id}/resolve:
    post:
      operationId: resolveManualFollowup
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Resolved, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
  /admin/support-issues/{issueId}/resolve:
    post:
      operationId: resolveSupportIssue
      tags: [admin]
      parameters: [{ name: issueId, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Resolved, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
  /admin/partners/{id}/archive:
    post:
      operationId: archivePartner
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Archived, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
  /admin/partners/{id}/unarchive:
    post:
      operationId: unarchivePartner
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Unarchived, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
  /admin/assets/{id}/archive:
    post:
      operationId: archiveAsset
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Archived, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
  /admin/assets/{id}/unarchive:
    post:
      operationId: unarchiveAsset
      tags: [admin]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        "200": { description: Unarchived, content: { application/json: { schema: { $ref: "#/components/schemas/AdminGenericObject" } } } }
...
components:
  schemas:
...
    AdminGenericObject:
      type: object
      additionalProperties: true
...